Remove Collector::get_epoch()
