+++ src/collector.rs
@@ -55,12 +55,6 @@ impl Collector {
         Handle::new(&self.0)
     }
 
-    /// Get the global epoch.
-    #[inline]
-    pub fn get_epoch(&self) -> usize {
-        self.0.get_epoch()
-    }
-
     /// Collect several bags from the global garbage queue and destroy their objects.
     ///
     /// # Safety
@@ -74,6 +68,9 @@ impl Collector {
 }
 
 impl Global {
+    /// Number of bags to destroy.
+    const COLLECT_STEPS: usize = 8;
+
     fn new() -> Self {
         Self {
             registries: List::new(),
@@ -81,9 +78,6 @@ impl Global {
             epoch: Epoch::new(),
         }
     }
-
-    /// Number of bags to destroy.
-    const COLLECT_STEPS: usize = 8;
 
     #[inline]
     pub fn get_epoch(&self) -> usize {
@@ -132,6 +126,44 @@ impl Global {
             unprotected(|scope| {
                 self.registries.insert(LocalEpoch::new(), scope).as_raw()
             })
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crossbeam_utils::scoped;
+
+    use super::*;
+
+    const NUM_THREADS: usize = 8;
+
+    #[test]
+    fn pin_holds_advance() {
+        let collector = Collector::new();
+
+        let threads = (0..NUM_THREADS)
+            .map(|_| {
+                scoped::scope(|scope| {
+                    scope.spawn(|| for _ in 0..100_000 {
+                        let handle = collector.handle();
+                        handle.pin(|_| {
+                            let before = collector.0.get_epoch();
+                            unsafe {
+                                collector.collect();
+                            }
+                            let after = collector.0.get_epoch();
+
+                            assert!(after.wrapping_sub(before) <= 2);
+                        });
+                    })
+                })
+            })
+            .collect::<Vec<_>>();
+        drop(collector);
+
+        for t in threads {
+            t.join();
         }
     }
 }
